==> 2012to2017/1. Get CRA data from FFIEC.ipynb <==
# Import packages, define directories to look at

print('Importing packages...')
import os
import urllib
import pandas
import zipfile
import time
#import re
from IPython.display import display, HTML

thisdir = '/home/idies/workspace/Storage/raddick/Baltimore/community_reinvestment_act/'
data_dir = '/home/idies/workspace/Temporary/raddick/cra_scratch/'
code_lookup_dir = thisdir + 'code_guide_lookups/'
#baltimore_dir = thisdir + 'baltimore/'

g = 0  # keep track of grand total of processing time

os.chdir(data_dir)
print('Now in directory: {0:}'.format(os.getcwd()))
print('ok')

# Get flatfiles for disclosure reports

## Download zipfiles of disclosure reports from ffiec.gov

# Get oroginal datafiles from ffiec.gov
s = time.time()

print('Downloading disclousre flatfiles from ffiec.gov...')
thatpath = 'https://www.ffiec.gov/cra/xls/'
theyears = list(range(96,100))
theyears += list(range(0,20))
filenames = []

for i in theyears:
    filenames.append('{:02d}exp_discl.zip'.format(i))
#filenames

for thisfile in filenames:
    print('Starting to download {:}...'.format(thisfile))
    with urllib.request.urlopen(thatpath+thisfile) as response:    
        it = response.read()
        with open(thisfile, 'wb') as f:
            f.write(it)
            
e = time.time()
g = g + (e-s)
print('Got {0:,.0f} files in {1:,.0f} seconds!'.format(len(filenames), e-s))


## Unzip disclosure flatfiles

#theyears = list(range(96,100))
#theyears += list(range(0,18))

s = time.time()
for i in theyears:
    thisfile = '{0:02d}exp_discl.zip'.format(i)
    if (thisfile in os.listdir()):
        print('Extracting {0:}...'.format(thisfile))
    else:
        print(thisfile+' not found!!!')
    thezipfile = zipfile.ZipFile(thisfile)
    thezipfile.extractall()
    thezipfile.close()
    if ((i >= 8) & (i <= 15)):
        os.rename('exp_discl.dat', '{0:02d}exp_discl.dat'.format(i))

print('Deleting zipfiles...')
thezipfiles = [x for x in os.listdir() if '.zip' in x]
for thiszip in thezipfiles:
    os.remove(thiszip)
e = time.time()
g = g + (e-s)
print('Extracted {0:,.0f} zipfiles in {1:,.0f} seconds!'.format(len(thezipfiles), e-s))


# Parse the single rawdata string into separate columns of strings

Using the guides on their website (https://www.ffiec.gov/cra/pdf/17FlatDiscSpecs.pdf)

Pre-2004 guides: https://www.ffiec.gov/cra/pdf/2003specs.pdf

## 2016-2019 load strings from multiple files into a DataFrame

s = time.time()
rdf = pandas.DataFrame()

minyear = 2016
maxyear = 2019
for y in range(maxyear, minyear-1, -1):
    for i in range(1,3):
        for j in range(1,3):
            thisfile = 'cra{0:04d}_Discl_D{1:.0f}{2:.0f}.dat'.format(y,i,j)
            print('Reading file {:}...'.format(thisfile))
            rdf = rdf.append(pandas.read_csv(thisfile, header=None))
    for k in range(3,7):
        thisfile = 'cra{0:04d}_Discl_D{1:.0f}.dat'.format(y,k)
        print('Reading file {:}...'.format(thisfile))
        rdf = rdf.append(pandas.read_csv(thisfile, header=None))
e = time.time()
g = g + (e-s)
print('\n')
print('Read {0:04d}-{1:04d} data: {2:,.0f} rows in {3:,.0f} seconds.'.format(minyear, maxyear, len(rdf), e-s))

## Now read 2004-2015 data (the long strings), append to the same DataFrame

s = time.time()

minyear = 2004
maxyear = 2015

filelist = []
for i in range(maxyear, minyear-1, -1):
    thisfile = '{:}exp_discl'.format(str(i)[2:])
    if i in [2004,2005,2006]:
        thisfile = thisfile + '_new.dat'
    else:
        thisfile = thisfile + '.dat'
    filelist.append(thisfile)
print('Found filenames!')
#print(filelist)
for thisfile in filelist:
    print('Reading file {:}...'.format(thisfile))
    rdf = rdf.append(pandas.read_csv(thisfile, header=None))

e = time.time()
g = g + (e-s)
print('Read {0:4d}-{1:04d} data: {2:,.0f} rows in {3:,.0f} seconds.'.format(minyear, maxyear, len(rdf), e-s))
print('Done!')

## Parse strings into DataFrame for 2004-2019

### Parse tableID so we can tell the difference between tables

Also get things that appear in the same place in every table: respondentID, agency_code, agency_year

s = time.time()
rdf.columns = ['thestring']
rdf.index.name = 'rownumber'
print('finding...')
print('tableID...')
rdf = rdf.assign(tableID = rdf['thestring'].apply(lambda x: x[0:5].strip()))
print('respondentID...')
rdf = rdf.assign(respondentID = pandas.to_numeric(rdf['thestring'].apply(lambda x: x[5:15]), errors='coerce'))
print('agency_code...')
rdf = rdf.assign(agency_code = pandas.to_numeric(rdf['thestring'].apply(lambda x: x[15]), errors='coerce'))
print('activity_year...')
rdf = rdf.assign(activity_year = pandas.to_numeric(rdf['thestring'].apply(lambda x: x[16:20]), errors='coerce'))
e = time.time()
g = g + (e-s)
print('Parsed {0:,.0f} rows in {1:,.0f} seconds!'.format(len(rdf), e-s))


### Use tableID to identify which table is which

s = time.time()

print('identifying...')
print('loans...')
loans_df = rdf[rdf['tableID'].apply(lambda x: x in ['D1-1', 'D1-2', 'D2-1', 'D2-2'])]
print('assessment areas...')
assessment_areas_df = rdf[rdf['tableID'].apply(lambda x: x in ['D3-0', 'D4-0'])]
print('community development loans...')
community_development_df = rdf[rdf['tableID'] == 'D5-0']
print('census tracts...')
tracts_df = rdf[rdf['tableID'] == 'D6-0']
e = time.time()
g = g + (e-s)
print('Done in {0:,.0f} seconds!'.format(e-s))


### Parse loans data

s = time.time()
print('Initially {0:,.0f} rows...'.format(len(loans_df)))
#s = time.time()
print('finding report_level...')
loans_df = loans_df.assign(report_level = loans_df['thestring'].apply(lambda x: x[42:45]))

print('keeping only lowest level of aggregation...')
loans_df = loans_df[loans_df['report_level'] == '   ']
print('Reduced to {0:,.0f} rows...'.format(len(loans_df)))

print('\nCreating new columns...')
print('loan_type...')
loans_df = loans_df.assign(loan_type = loans_df['thestring'].apply(lambda x: x[20]))
print('action_taken_type...')
loans_df = loans_df.assign(action_taken_type = loans_df['thestring'].apply(lambda x: x[21]))
print('state...')
loans_df = loans_df.assign(state = loans_df['thestring'].apply(lambda x: x[22:24]))
print('county...')
loans_df = loans_df.assign(county = loans_df['thestring'].apply(lambda x: x[24:27]))
print('msa...')
loans_df = loans_df.assign(msa = loans_df['thestring'].apply(lambda x: x[27:32]))
print('assessment_area_number...')
loans_df = loans_df.assign(assessment_area_number = loans_df['thestring'].apply(lambda x: x[32:36]))
print('partial_county_indicator...')
loans_df = loans_df.assign(partial_county_indicator = loans_df['thestring'].apply(lambda x: x[36]))
print('split_county_indicator...')
loans_df = loans_df.assign(split_county_indicator = loans_df['thestring'].apply(lambda x: x[37]))
print('population_classification...')
loans_df = loans_df.assign(population_classification = loans_df['thestring'].apply(lambda x: x[38]))
print('income_group_total...')
loans_df = loans_df.assign(income_group_total = loans_df['thestring'].apply(lambda x: x[39:42]))
print('nLoans1...')
loans_df = loans_df.assign(nLoans1 = loans_df['thestring'].apply(lambda x: x[45:55]))
print('amtLoans1...')
loans_df = loans_df.assign(amtLoans1 = loans_df['thestring'].apply(lambda x: x[55:65]))
print('nLoans100k...')
loans_df = loans_df.assign(nLoans100k = loans_df['thestring'].apply(lambda x: x[65:75]))
print('amtLoans100k...')
loans_df = loans_df.assign(amtLoans100k = loans_df['thestring'].apply(lambda x: x[75:85]))
print('nLoans250k...')
loans_df = loans_df.assign(nLoans250k = loans_df['thestring'].apply(lambda x: x[85:95]))
print('amtLoans250k...')
loans_df = loans_df.assign(amtLoans250k = loans_df['thestring'].apply(lambda x: x[95:105]))
print('nLoansToSmallest...')
loans_df = loans_df.assign(nLoansToSmallest = loans_df['thestring'].apply(lambda x: x[105:115]))
print('amtLoansToSmallest...')
loans_df = loans_df.assign(amtLoansToSmallest = loans_df['thestring'].apply(lambda x: x[115:125]))
print('nLoansAff...')
loans_df = loans_df.assign(nLoansAff = loans_df['thestring'].apply(lambda x: x[125:135]))
print('amtLoansAff..\n')
loans_df = loans_df.assign(amtLoansAff = loans_df['thestring'].apply(lambda x: x[135:145]))

numeric_columns = ['loan_type','action_taken_type','state','county','msa']
numeric_columns += ['assessment_area_number','income_group_total']
numeric_columns += ['nLoans1','amtLoans1','nLoans100k','amtLoans100k']
numeric_columns += ['nLoans250k','amtLoans250k','nLoansToSmallest','amtLoansToSmallest']
numeric_columns += ['nLoansAff','amtLoansAff']

print('\nConverting to numeric...')
for x in numeric_columns:
    print('{0:}...'.format(x))
    loans_df.loc[:, x] = pandas.to_numeric(loans_df[x], errors='coerce')

print('Multiplying loan amounts by 1000 for real dollars...')
money_columns = ['amtLoans1', 'amtLoans100k', 'amtLoans250k']
money_columns += ['amtLoansToSmallest', 'amtLoansAff']
for x in money_columns:
    loans_df.loc[:, x] = loans_df[x] * 1000

print('Dropping columns we do not need...')
loans_df = loans_df.drop(['thestring', 'tableID', 'report_level'], axis=1)
e = time.time()
g = g + (e-s)

print('Writing out loans data...')
loans_df.to_csv('loans.csv', encoding='utf-8')

print('\nProcessed {0:,.0f} rows in {1:,.0f} seconds!'.format(len(loans_df), e-s))


### Parse assessment areas data

s = time.time()
print('finding report level...')
assessment_areas_df = assessment_areas_df.assign(report_level = assessment_areas_df['thestring'].apply(lambda x: x[37:39]))

print('keeping only report_level = 10 (so we can connect to county)...')
assessment_areas_df = assessment_areas_df[assessment_areas_df['report_level'] == '10']

print('loan_type...')
assessment_areas_df = assessment_areas_df.assign(loan_type = assessment_areas_df['thestring'].apply(lambda x: x[20]))
print('state...')
assessment_areas_df = assessment_areas_df.assign(state = assessment_areas_df['thestring'].apply(lambda x: x[21:23]))
print('county...')
assessment_areas_df = assessment_areas_df.assign(county = assessment_areas_df['thestring'].apply(lambda x: x[23:26]))
print('msa...')
assessment_areas_df = assessment_areas_df.assign(msa = assessment_areas_df['thestring'].apply(lambda x: x[26:31]))
print('assessment_area_number...')
assessment_areas_df = assessment_areas_df.assign(assessment_area_number = assessment_areas_df['thestring'].apply(lambda x: x[31:35]))
print('partial_county_indicator...')
assessment_areas_df = assessment_areas_df.assign(partial_county_indicator = assessment_areas_df['thestring'].apply(lambda x: x[35]))
print('split_county_indicator...')
assessment_areas_df = assessment_areas_df.assign(split_county_indicator = assessment_areas_df['thestring'].apply(lambda x: x[36]))

print('nBizLoans...')
assessment_areas_df = assessment_areas_df.assign(nBizLoans = assessment_areas_df['thestring'].apply(lambda x: x[39:49]))
print('amtBizLoans...')
assessment_areas_df = assessment_areas_df.assign(amtBizLoans = assessment_areas_df['thestring'].apply(lambda x: x[49:59]))
print('nLoansToSmallest...')
assessment_areas_df = assessment_areas_df.assign(nLoansToSmallest = assessment_areas_df['thestring'].apply(lambda x: x[59:69]))
print('amtLoansToSmallest...')
assessment_areas_df = assessment_areas_df.assign(amtLoansToSmallest = assessment_areas_df['thestring'].apply(lambda x: x[69:79]))
print('nLoansPurchased...')
assessment_areas_df = assessment_areas_df.assign(nLoansPurchased = assessment_areas_df['thestring'].apply(lambda x: x[79:89]))
print('amtLoansPurchased...')
assessment_areas_df = assessment_areas_df.assign(amtLoansPurchased = assessment_areas_df['thestring'].apply(lambda x: x[89:99]))
#assessment_areas_df.head(10)
print('dropping columns we do not need...')
assessment_areas_df = assessment_areas_df.drop(['thestring', 'tableID', 'report_level'], axis=1)
assessment_areas_df.head(1).T

print('converting to numeric...')
numeric_columns = ['respondentID','agency_code','activity_year']
numeric_columns += ['loan_type','state','county','msa','assessment_area_number']
numeric_columns += ['nBizLoans','amtBizLoans','nLoansToSmallest','amtLoansToSmallest']
numeric_columns += ['amtLoansPurchased','nLoansPurchased']
for x in numeric_columns:
    assessment_areas_df.loc[:, x] = pandas.to_numeric(assessment_areas_df[x], errors='coerce')

print('Multiplying loan amounts by 1000 for real dollars...')
money_columns = ['amtBizLoans', 'amtLoansToSmallest', 'amtLoansPurchased']
for x in money_columns:
    assessment_areas_df.loc[:, x] = assessment_areas_df[x] * 1000

print('Writing out...')
assessment_areas_df.to_csv('assessment_areas.csv', encoding='utf-8')

e = time.time()
g = g + (e-s)

print('Processed {0:,.0f} rows in {1:,.0f} seconds!'.format(len(assessment_areas_df), e-s))


### Parse community development and third-party loans data

print('Getting community development and third-party loans...')
s = time.time()
community_development_df = community_development_df.assign(loan_type = community_development_df['thestring'].apply(lambda x: x[20]))
community_development_df = community_development_df.assign(nLoans = community_development_df['thestring'].apply(lambda x: x[21:31]))
community_development_df = community_development_df.assign(amtLoans = community_development_df['thestring'].apply(lambda x: x[31:41]))
community_development_df = community_development_df.assign(nLoansAff = community_development_df['thestring'].apply(lambda x: x[41:51]))
community_development_df = community_development_df.assign(amtLoansAff = community_development_df['thestring'].apply(lambda x: x[51:61]))
community_development_df = community_development_df.assign(action_taken_type = community_development_df['thestring'].apply(lambda x: x[61]))

community_development_df = community_development_df.drop(['thestring', 'tableID'], axis=1)

numeric_columns = ['loan_type','nLoans','amtLoans','nLoansAff','amtLoansAff']
for x in numeric_columns:
    community_development_df.loc[:, x] = pandas.to_numeric(community_development_df[x], errors='coerce')

money_columns = ['amtLoans', 'amtLoansAff']
for x in money_columns:
    community_development_df.loc[:, x] = community_development_df[x] * 1000

print('Writing out...')
community_development_df.to_csv('community_development.csv', encoding='utf-8')

e = time.time()
g = g + (e-s)

print('Read {0:,.0f} rows in {1:,.0f} seconds.'.format(len(community_development_df), e-s))


### Parse census tracts data

s = time.time()
print('assigning...')
print('state...')
tracts_df = tracts_df.assign(state = tracts_df['thestring'].apply(lambda x: x[20:22]))
print('county...')
tracts_df = tracts_df.assign(county = tracts_df['thestring'].apply(lambda x: x[22:25]))
print('msa...')
tracts_df = tracts_df.assign(msa = tracts_df['thestring'].apply(lambda x: x[25:30]))
print('census_tract...')
tracts_df = tracts_df.assign(census_tract = tracts_df['thestring'].apply(lambda x: x[30:37]))
print('assessment_area_number...')
tracts_df = tracts_df.assign(assessment_area_number = tracts_df['thestring'].apply(lambda x: x[37:41]))
print('partial_county_indicator...')
tracts_df = tracts_df.assign(partial_county_indicator = tracts_df['thestring'].apply(lambda x: x[41]))
print('split_county_indicator...')
tracts_df = tracts_df.assign(split_county_indicator = tracts_df['thestring'].apply(lambda x: x[42]))
print('population_classification...')
tracts_df = tracts_df.assign(population_classification = tracts_df['thestring'].apply(lambda x: x[43]))
print('income_group...')
tracts_df = tracts_df.assign(income_group = tracts_df['thestring'].apply(lambda x: x[44:47]))
print('loan_indicator...')
tracts_df = tracts_df.assign(loan_indicator = tracts_df['thestring'].apply(lambda x: x[47]))

print('\ndropping columns we do not need...')
tracts_df = tracts_df.drop(['thestring', 'tableID'], axis=1)

print('\nconverting to numeric...')
numeric_columns = ['state', 'county', 'msa', 'census_tract']
numeric_columns += ['assessment_area_number', 'income_group']
for x in numeric_columns:
    print('{0:}...'.format(x))
    tracts_df.loc[:, x] = pandas.to_numeric(tracts_df[x], errors='coerce')

e = time.time()
g = g + (e-s)

print('Writing out...')
tracts_df.to_csv('tracts.csv', encoding='utf-8')

print('Processed {0:,.0f} rows in {1:,.0f} seconds!'.format(len(tracts_df), e-s))

##agency_code
#code,value
#1,'OCC'
#2,'FRS'
#3,'FDIC'
#4,'OTS'

##loan_type
#3,'small business'
#4,'small farm'

##action_taken_type
#1,'origination'
#6,'purchase'

### Delete .dat files

print('Removing .dat files...')
deletefiles = [data_dir+x for x in os.listdir(data_dir) if '.dat' in x]
for thisfile in deletefiles:
    os.remove(thisfile)
print('Done!')


# Get aggregate data

File specs: https://www.ffiec.gov/cra/pdf/17FlatAggSpecs.pdf

## Get zipfiles from ffiec.gov

# Get aggregate datafiles from ffiec.gov
s = time.time()
print('Downloading aggregate data flatfiles...')
thatpath = 'https://www.ffiec.gov/cra/xls/'
theyears = list(range(96,100))
theyears += list(range(0,20))


filenames = []

for i in theyears:
    filenames.append('{:02d}exp_aggr.zip'.format(i))
#filenames

for thisfile in filenames:
    print('Starting to download {:}...'.format(thisfile))
    with urllib.request.urlopen(thatpath+thisfile) as response:    
        it = response.read()
        with open(thisfile, 'wb') as f:
            f.write(it)
            
e = time.time()
g = g + (e-s)
print('Got {0:,.0f} files in {1:,.0f} seconds!'.format(len(filenames), e-s))


## Unzip

#theyears = list(range(96,100))
#theyears += list(range(0,18))
#theyears = [17]
s = time.time()

for i in theyears:
    thisfile = '{0:02d}exp_aggr.zip'.format(i)
    if (thisfile in os.listdir()):
        print('Extracting {0:}...'.format(thisfile))
    else:
        print(thisfile+' not found!!!')
    thezipfile = zipfile.ZipFile(thisfile)
    thezipfile.extractall()
    thezipfile.close()
    if (i in (5,6,8,9,10,11,12,13,14,15)):
        os.rename('exp_aggr.dat', '{0:02d}exp_aggr.dat'.format(i))

print('Deleting zipfiles...')
thezipfiles = [x for x in os.listdir() if '.zip' in x]
for thiszip in thezipfiles:
    os.remove(thiszip)
e = time.time()
g = g + (e-s)
print('Extracted {0:,.0f} zipfiles in {1:,.0f} seconds!'.format(len(theyears), e-s))


## 2016-2019: load strings from multiple files into a DataFrame

s = time.time()
agg_loans_df = pandas.DataFrame()

minyear = 2016
maxyear = 2019
for y in range(maxyear, minyear-1, -1):
    for i in range(1,3):
        for j in range(1,3):
            thisfile = 'cra{0:04d}_Aggr_A{1:.0f}{2:.0f}.dat'.format(y,i,j)
            print('Reading file {:}...'.format(thisfile))
            agg_loans_df = agg_loans_df.append(pandas.read_csv(thisfile, header=None))
e = time.time()
g = g + (e-s)
print('\n')
print('Read {0:04d}-{1:04d} data: {2:,.0f} rows in {3:,.0f} seconds.'.format(minyear, maxyear, len(agg_loans_df), e-s))


## Now read 2004-2015 data (the long strings), append to the same DataFrame

s = time.time()

minyear = 2004
maxyear = 2015

filelist = []
for i in range(maxyear, minyear-1, -1):
    thisfile = '{:}exp_aggr.dat'.format(str(i)[2:])
    filelist.append(thisfile)
print('Found filenames!')
#print(filelist)
for thisfile in filelist:
    print('Reading file {:}...'.format(thisfile))
    agg_loans_df = agg_loans_df.append(pandas.read_csv(thisfile, header=None))
e = time.time()
g = g + (e-s)
print('Read {0:4d}-{1:04d} data: {2:,.0f} rows in {3:,.0f} seconds.'.format(minyear, maxyear, len(agg_loans_df), e-s))


### Use tableID to keep only aggregated loans data (not lenders)

### Parse aggregate loans data

s = time.time()
agg_loans_df.columns = ['thestring']
agg_loans_df.index.name = 'rownumber'

print('Assigning tableID...')
agg_loans_df = agg_loans_df.assign(tableID = agg_loans_df['thestring'].apply(lambda x: x[0:5]))
print('keeping only loans data...')
agg_loans_df = agg_loans_df[agg_loans_df['tableID'].apply(lambda x: 'a' not in x)]

print('Assigning report_level...')
agg_loans_df = agg_loans_df.assign(report_level = agg_loans_df['thestring'].apply(lambda x: x[33:36]))
print('keeping only base level data (report_level blank)...')
agg_loans_df = agg_loans_df[agg_loans_df['report_level'] == '   ']

print('\nAssigning...')
print('activity_year...')
agg_loans_df = agg_loans_df.assign(activity_year = agg_loans_df['thestring'].apply(lambda x: x[5:9]))
print('loan_type...')
agg_loans_df = agg_loans_df.assign(loan_type = agg_loans_df['thestring'].apply(lambda x: x[9]))
print('action_taken_type...')
agg_loans_df = agg_loans_df.assign(action_taken_type = agg_loans_df['thestring'].apply(lambda x: x[10]))
print('state...')
agg_loans_df = agg_loans_df.assign(state = agg_loans_df['thestring'].apply(lambda x: x[11:13]))
print('county...')
agg_loans_df = agg_loans_df.assign(county = agg_loans_df['thestring'].apply(lambda x: x[13:16]))
print('msa...')
agg_loans_df = agg_loans_df.assign(msa = agg_loans_df['thestring'].apply(lambda x: x[16:21]))
print('census_tract...')
agg_loans_df = agg_loans_df.assign(census_tract = agg_loans_df['thestring'].apply(lambda x: x[21:28]))
print('split_county_indicator...')
agg_loans_df = agg_loans_df.assign(split_county_indicator = agg_loans_df['thestring'].apply(lambda x: x[28]))
print('population_classification...')
agg_loans_df = agg_loans_df.assign(population_classification = agg_loans_df['thestring'].apply(lambda x: x[29]))
print('income_group_total...')
agg_loans_df = agg_loans_df.assign(income_group_total = agg_loans_df['thestring'].apply(lambda x: x[30:33]))
print('nLoans1...')
agg_loans_df = agg_loans_df.assign(nLoans1 = agg_loans_df['thestring'].apply(lambda x: x[36:46]))
print('amtLoans1...')
agg_loans_df = agg_loans_df.assign(amtLoans1 = agg_loans_df['thestring'].apply(lambda x: x[46:56]))
print('nLoans100k...')
agg_loans_df = agg_loans_df.assign(nLoans100k = agg_loans_df['thestring'].apply(lambda x: x[56:66]))
print('amtLoans100k...')
agg_loans_df = agg_loans_df.assign(amtLoans100k = agg_loans_df['thestring'].apply(lambda x: x[66:76]))
print('nLoans250k...')
agg_loans_df = agg_loans_df.assign(nLoans250k = agg_loans_df['thestring'].apply(lambda x: x[76:86]))
print('amtLoans250k...')
agg_loans_df = agg_loans_df.assign(amtLoans250k = agg_loans_df['thestring'].apply(lambda x: x[86:96]))
print('nLoansToSmalest...')
agg_loans_df = agg_loans_df.assign(nLoansToSmallest = agg_loans_df['thestring'].apply(lambda x: x[96:106]))
print('amtLoansToSmalest...')
agg_loans_df = agg_loans_df.assign(amtLoansToSmallest = agg_loans_df['thestring'].apply(lambda x: x[106:116]))

print('\nDeleting columns we do not need...')
agg_loans_df = agg_loans_df.drop(['thestring', 'tableID', 'report_level'], axis=1)

print('Converting columns to numeric...')
numeric_columns = ['activity_year', 'loan_type', 'action_taken_type']
numeric_columns += ['state', 'county', 'msa', 'census_tract']
numeric_columns += ['income_group_total', 'nLoans1', 'amtLoans1']
numeric_columns += ['nLoans100k', 'amtLoans100k']
numeric_columns += ['nLoans250k', 'amtLoans250k']
numeric_columns += ['nLoansToSmallest', 'amtLoansToSmallest']

for x in numeric_columns:
    print('{0:}...'.format(x))
    agg_loans_df.loc[:, x] = pandas.to_numeric(agg_loans_df[x], errors='coerce')

print('Multiplying loan amounts by 1000 for real dollars...')
money_columns = ['amtLoans1', 'amtLoans100k', 'amtLoans250k']
money_columns += ['amtLoansToSmallest']
for x in money_columns:
    agg_loans_df.loc[:, x] = agg_loans_df[x] * 1000

print('Writing out...')
agg_loans_df.to_csv('agg_loans.csv', encoding='utf-8')
e = time.time()
g = g + (e-s)

print('Processed {0:,.0f} rows in {1:,.0f} seconds!'.format(len(agg_loans_df), e-s))

#agg_loans_df.head(1).T

## Parse aggregated lender data

s = time.time()
agg_lenders_df = pandas.DataFrame()

minyear = 2016
maxyear = 2019
for y in range(maxyear, minyear-1, -1):
    for i in range(1,3):
        for j in range(1,3):
            thisfile = 'cra{0:04d}_Aggr_A{1:.0f}{2:.0f}a.dat'.format(y,i,j)
            print('Reading file {:}...'.format(thisfile))
            agg_lenders_df = agg_lenders_df.append(pandas.read_csv(thisfile, header=None))
e = time.time()
g = g + (e-s)
print('Read {0:04d}-{1:04d} data: {2:,.0f} rows...'.format(minyear, maxyear, len(agg_lenders_df)))

minyear = 2004
maxyear = 2015

filelist = []
for i in range(maxyear, minyear-1, -1):
    thisfile = '{:}exp_aggr.dat'.format(str(i)[2:])
    filelist.append(thisfile)
print('\n')
print('Found filenames!')
#print(filelist)
for thisfile in filelist:
    print('Reading file {:}...'.format(thisfile))
    agg_lenders_df = agg_lenders_df.append(pandas.read_csv(thisfile, header=None))
e = time.time()
g = g + (e-s)
print('Read {0:4d}-{1:04d} data: {2:,.0f} rows...'.format(minyear, maxyear, len(agg_loans_df), e-s))
print('\n')

agg_lenders_df.columns = ['thestring']
agg_lenders_df.index.name = 'rownumber'

print('Assigning tableID...')
agg_lenders_df = agg_lenders_df.assign(tableID = agg_lenders_df['thestring'].apply(lambda x: x[0:5]))
print('keeping only loans data...')
agg_lenders_df = agg_lenders_df[agg_lenders_df['tableID'].apply(lambda x: 'a' in x)]

print('Assigning report_level...')
agg_lenders_df = agg_lenders_df.assign(report_level = agg_lenders_df['thestring'].apply(lambda x: x[37:40]))
print('keeping only lenders data...')
agg_lenders_df = agg_lenders_df[agg_lenders_df['report_level'] == '   ']

print('\nAssigning...')
print('activity_year...')
agg_lenders_df = agg_lenders_df.assign(activity_year = agg_lenders_df['thestring'].apply(lambda x: x[5:9]))
print('loan_type...')
agg_lenders_df = agg_lenders_df.assign(loan_type = agg_lenders_df['thestring'].apply(lambda x: x[9]))
print('action_taken_type...')
agg_lenders_df = agg_lenders_df.assign(action_taken_type = agg_lenders_df['thestring'].apply(lambda x: x[10]))
print('state...')
agg_lenders_df = agg_lenders_df.assign(state = agg_lenders_df['thestring'].apply(lambda x: x[11:13]))
print('county...')
agg_lenders_df = agg_lenders_df.assign(county = agg_lenders_df['thestring'].apply(lambda x: x[13:16]))
print('msa...')
agg_lenders_df = agg_lenders_df.assign(msa = agg_lenders_df['thestring'].apply(lambda x: x[16:21]))
print('respondentID...')
agg_lenders_df = agg_lenders_df.assign(respondentID = agg_lenders_df['thestring'].apply(lambda x: x[21:31]))
print('agency_code...')
agg_lenders_df = agg_lenders_df.assign(agency_code = agg_lenders_df['thestring'].apply(lambda x: x[31]))
print('nLenders...')
agg_lenders_df = agg_lenders_df.assign(nLenders = agg_lenders_df['thestring'].apply(lambda x: x[32:37]))
print('nLoans...')
agg_lenders_df = agg_lenders_df.assign(nLoans = agg_lenders_df['thestring'].apply(lambda x: x[40:50]))
print('amtLoans...')
agg_lenders_df = agg_lenders_df.assign(amtLoans = agg_lenders_df['thestring'].apply(lambda x: x[50:60]))
print('nLoansToSmallest...')
agg_lenders_df = agg_lenders_df.assign(nLoansToSmallest = agg_lenders_df['thestring'].apply(lambda x: x[60:70]))
print('amtLoansToSmallest...')
agg_lenders_df = agg_lenders_df.assign(amtLoansToSmallest = agg_lenders_df['thestring'].apply(lambda x: x[70:80]))

print('\nDropping columns we do not need...')
agg_lenders_df = agg_lenders_df.drop(['thestring', 'tableID', 'report_level'], axis=1)

numeric_columns = ['activity_year', 'loan_type', 'action_taken_type']
numeric_columns += ['state', 'county', 'msa', 'respondentID', 'agency_code']
numeric_columns += ['nLenders', 'nLoans', 'amtLoans']
numeric_columns += ['nLoansToSmallest', 'amtLoansToSmallest']

for x in numeric_columns:
    agg_lenders_df.loc[:, x] = pandas.to_numeric(agg_lenders_df[x], errors='coerce')
    
print('Multiplying loan amounts by 1000 for real dollars...')
money_columns = ['amtLoans', 'amtLoansToSmallest']
for x in money_columns:
    agg_lenders_df.loc[:, x] = agg_lenders_df[x] * 1000

print('Writing out...')
agg_lenders_df.to_csv('agg_lenders.csv', encoding='utf-8')

e = time.time()
g = g + (e-s)
print('Processed {0:,.0f} rows in {1:,.0f} seconds!'.format(len(agg_lenders_df), e-s))



### Delete .dat files

print('Removing .dat files...')
deletefiles = [data_dir+x for x in os.listdir(data_dir) if '.dat' in x]
for thisfile in deletefiles:
    os.remove(thisfile)
print('Done!')
