==> 2012to2017/2. Get LODES WAC data.ipynb <==
import os
from urllib import request
from bs4 import BeautifulSoup 
import time
import gzip
import numpy as np
import pandas
from pprint import pprint
#import io
data_dir = '/home/idies/workspace/Temporary/raddick/cra_scratch/'
jobs_dir = data_dir + 'lodes_wac/'
baltimore_dir = '/home/idies/workspace/Storage/raddick/Baltimore/community_reinvestment_act/baltimore/'

this_state = 'wy'
this_state_i = 62


g = 0
print('Done!')
os.getcwd()

s = time.time()
y = os.listdir(data_dir)
hasdir = [x for x in y if x=='lodes_wac']
if (len(hasdir) == 0):
    os.mkdir(jobs_dir)
e = time.time()
g = g + (e-s)
print('Done in {0:,.3f} seconds!'.format(e-s))
#os.listdir(jobs_dir)
os.listdir(jobs_dir)

# Get WAC data from census.gov and combine into one file per state

s = time.time()


theurl = 'https://lehd.ces.census.gov/data/lodes/LODES7/'
print('scraping data...')
page = request.urlopen(theurl)
soup = BeautifulSoup(page, 'html.parser')
page.close()

bigtable = soup.find('table')
allstaterows = bigtable.findAll('tr')

#for i in range(9,10): # Alaska: wrote 1,975,372 rows to jobs_data_ak.csv in 24 minutes 40 seconds!
#for i in range(10,11): # Alabama:
#for i in range(11,12): # Arkansas: wrote 11,346,396 rows to jobs_data_ar.csv in 72 minutes 26 seconds!
#for i in range(12,13): # Arizona: wrote 14,821,818 rows to jobs_data_az.csv in 90 minutes 31 seconds!

#for i in range(13,14): # California 0, California 1
# California 2: wrote 24,754,939 rows to jobs_data_ca_2.csv in 46 minutes 47 seconds!
# California 3: Wrote 20,696,087 rows to jobs_data_ca_3.csv in 44 minutes 35 seconds!
# Colorado: 
# Connecticut: wrote 11,443,174 rows to jobs_data_ct.csv in 79 minutes 1 seconds!
# DC: wrote 824,450 rows to jobs_data_dc.csv in 5 minutes 21 seconds!
# Delaware: wrote 2,963,580 rows to jobs_data_de.csv in 23 minutes 15 seconds!

for i in range(this_state_i, this_state_i + 1):
    allstatecells = allstaterows[i].findAll('td')
    this_state_url = theurl + allstatecells[1].find('a').get('href')+'wac/'
    print('Scraping from: {0:}'.format(this_state_url))
    #print(this_state_url)
    statepage = request.urlopen(this_state_url)
    statesoup = BeautifulSoup(statepage, 'html.parser')
    statepage.close()
    statetable = statesoup.find('table')
    filerows = statetable.findAll('tr')
    print('Parsing state {0:,.0f} of {1:,.0f}... Getting {2:,.0f} files...'.format(i, 60, len(filerows)-4))
    for j in range(3, len(filerows)-1):
        if ((np.mod(j, 100) == 0) | (j == len(filerows)-2)):
            print('\tReading file {0:,.0f} of {1:,.0f}...'.format(j, len(filerows)-4))
        filecells = filerows[j].findAll('td')
        this_file_url = this_state_url + filecells[1].find('a').get('href')
        this_file_name = jobs_dir + filecells[1].text
        with open(this_file_name, 'wb') as f:
            r = request.urlopen(this_file_url)
            f.write(r.read())
            r.close()
nFiles = len([x for x in os.listdir(jobs_dir) if '.gz' in x])
e = time.time()
g = g + (e-s)
print('Scrapered data to create {0:,.0f} files in {1:,.0f} minutes {2:,.0f} seconds!'.format(nFiles, np.floor((e-s)/60), (e-s)%60))


### Unzip files

s = time.time()

print('Unzipping files...')
gzipfiles = [jobs_dir+x for x in os.listdir(jobs_dir) if ('{0:}_wac'.format(this_state) in x) and (x[-3:] == '.gz')]
print('Found {0:,.0f} gzip files...'.format(len(gzipfiles)))

for i in range(0, len(gzipfiles)):
    if ((np.mod(i, 100) == 0) or (i == len(gzipfiles) - 1)):
        print('\tUn-gzipping file {0:.0f} of {1:.0f}...'.format(i, len(gzipfiles)-1))
    csvfilename = gzipfiles[i][:-3]
    with gzip.open(gzipfiles[i], 'rb') as f:
        file_content = f.read()
    with open(csvfilename, 'wb') as f:
        f.write(file_content)
e = time.time()
g = g + (e-s)
print('Files unzipped in {0:,.0f} minutes {1:,.0f} seconds...'.format(np.floor((e-s)/60), (e-s) % 60))

s = time.time()
print('Deleting zipfiles...')
#os.listdir(jobs_dir)
gzfiles = [jobs_dir+x for x in os.listdir(jobs_dir) if x[-3:] == '.gz']
for y in gzfiles:
    os.remove(y)
e = time.time()
g = g + (e-s)

print('Deleted all .gz files in {0:,.0f} seconds...'.format(e-s))
#print('Done in {0:,.0f} seconds!'.format(g))
print('Done!')

s = time.time()

print('Combining multiple files into a single CSV for {0:}...'.format(this_state))
jobs_df = pandas.DataFrame()
infiles = [jobs_dir+x for x in os.listdir(jobs_dir) if ('{0:}_wac'.format(this_state) in x) and (x[-4:] == '.csv')]
outfiles = [jobs_dir+x for x in os.listdir(jobs_dir) if ('{0:}'.format(this_state) in x) and ('jobs_data_' in x)]

noutputfiles = len(outfiles)
#print(noutputfiles)

if (this_state in ['ca', 'fl', 'ga', 'il', 'ny', 'tx', 'va']):   # ['ca','fl', 'il', 'ny', 'tx', 'va']
    if (noutputfiles == 0):
        startfile = 0
        last_to_read = int(np.floor(len(infiles)/4))
        indexer = 0
    elif (noutputfiles == 1):
        startfile = int(np.floor(len(infiles)/4))
        last_to_read = int(np.floor(len(infiles)/2))
        indexer = 1
    elif (noutputfiles == 2):
        startfile = int(np.floor(len(infiles)/2))
        last_to_read = int(3*np.floor(len(infiles)/4))
        indexer = 2
    elif (noutputfiles == 3):
        startfile = int(3*np.floor(len(infiles)/4))
        last_to_read =len(infiles)
        indexer = 3
else:
    startfile = 0
    last_to_read = len(infiles)

for i in range(startfile, last_to_read):
    if ((np.mod(i, 25) == 0) or (i == len(infiles) - 1)):
        print('\tCombining file {0:,.0f} of {1:,.0f}...'.format(i, last_to_read-1))
    thisyear = int(infiles[i][-8:-4])
    #print(thisyear)
    xdf = pandas.read_csv(infiles[i], low_memory=False)
    xdf = xdf.assign(year = thisyear)     
    jobs_df = jobs_df.append(xdf)
jobs_df = jobs_df.reset_index(drop=True)
jobs_df.index.name = 'rownumber'

print('writing outfile...')

if (this_state in ['ca', 'fl', 'ga', 'il', 'ny', 'tx', 'va']): 
    jobs_df.to_csv(jobs_dir+'jobs_data_{0:}_{1:.0f}.csv'.format(this_state, indexer), encoding='utf-8')
else:
    jobs_df.to_csv(jobs_dir+'jobs_data_{0:}.csv'.format(this_state), encoding='utf-8')
e = time.time()
g = g + (e-s)

if (this_state in ['ca', 'fl', 'ga', 'il', 'ny', 'tx', 'va']):  
    print('Wrote {0:,.0f} rows to jobs_data_{1:}_{2:.0f}.csv in {3:,.0f} minutes {4:,.0f} seconds!\n'.format(len(jobs_df), this_state, indexer, np.floor((e-s)/60), (e-s) % 60))
else:
    print('Done: wrote {0:,.0f} rows to jobs_data_{1:}.csv in {2:,.0f} minutes {3:,.0f} seconds!\n'.format(len(jobs_df), this_state, np.floor((g)/60), (g) % 60))

# FL 2: Wrote 15,389,748 rows to jobs_data_fl_1.csv in 24 minutes 11 seconds!
# FL 3: Wrote 11,163,047 rows to jobs_data_fl_3.csv in 21 minutes 29 seconds!

# GA 0: Wrote 9,466,819 rows to jobs_data_ga_0.csv in 14 minutes 9 seconds!
# GA 1: Wrote 7,337,549 rows to jobs_data_ga_1.csv in 10 minutes 28 seconds!
# GA 2: Wrote 7,115,544 rows to jobs_data_ga_2.csv in 10 minutes 59 seconds!
# GA 3: Wrote 5,311,130 rows to jobs_data_ga_3.csv in 9 minutes 52 seconds!
# ID: wrote 8,393,228 rows to jobs_data_id.csv in 61 minutes 56 seconds!
# IL 0: Wrote 14,595,212 rows to jobs_data_il_0.csv in 21 minutes 52 seconds!


print('Deleting component files that we combined...')
s = time.time()
jobs_files_list = sorted([x[-6:-4] for x in os.listdir(jobs_dir) if 'jobs_' in x])
print('Found {0:.0f} state jobs files!'.format(len(jobs_files_list)))
#pprint(jobs_files_list)
files_to_delete = sorted([jobs_dir+y for y in os.listdir(jobs_dir) if ('.csv' in y) and ('jobs_' not in y)])
print('Will delete {0:,.0f} files...'.format(len(files_to_delete)))
print('Deleting...')
for z in files_to_delete:
    os.remove(z)
e = time.time()
g = g + (e-s)
print('Deleted files in {0:,.0f} minutes {1:,.0f} seconds!\n'.format(np.floor((e-s)/60), (e-s) % 60))



# this_state = 'il'
# i = 0
df = pandas.read_csv(jobs_dir+'jobs_data_{0:}.csv'.format(this_state))#, low_memory=False, index_col='rownumber', keep_default_na=False)
#df = pandas.read_csv(jobs_dir+'jobs_data_{0:}_{1:.0f}.csv'.format(this_state, i))#, low_memory=False, index_col='rownumber', keep_default_na=False)

df = df.set_index('rownumber')
#

print(df.groupby('year').size().sort_index())

df.sample(1).T


# Get Baltimore data and write it out

s = time.time()
df = pandas.read_csv(jobs_dir+'jobs_data_{0:}.csv'.format(this_state), encoding='utf-8', low_memory=False, index_col='rownumber')
#df = pandas.read_csv(jobs_dir+'jobs_data_{0:}.csv'.format(this_state), encoding='utf-8', low_memory=False, index_col='rownumber')
#baltimore_df = df[df['w_geocode'].apply(lambda x: str(x)[0:5] == '24510')]
#baltimore_df.to_csv(baltimore_dir+'wac_jobs_df.csv', encoding='utf-8')
e = time.time()
g = g + (e-s)
print('Read {0:,.0f} rows in {1:,.0f} seconds!'.format(len(df), e-s))
df.head(2)

df.tail(3)

s = time.time()
jobs_meta_df = pandas.read_csv(data_dir+'lodes_wac_codebook.csv', encoding='utf-8', low_memory=False, index_col='varnum')
jobs_meta_df.to_csv(baltimore_dir+'wac_jobs_metadata.csv', encoding='utf-8')
e = time.time()
g = g + (e-s)
print('Wrote metadata for {0:,.0f} variables in {1:,.0f} seconds!'.format(len(jobs_meta_df), e-s))
