==> 2012to2017/3. Read ACS5 metadata.ipynb <==
#!pip install --upgrade pip
#!pip install xlrd

import numpy as np
import pandas
import time
import pprint
import geopandas
import os
import re
#from SciServer import CasJobs
#from pprint import pprint

pandas.set_option('display.max_colwidth', None)
print('ok')

s = time.time()

debug = 1

for thisyear in range(2019,2009,-1):
    print('Processing metadata for {0:.0f}...'.format(thisyear))
    basedir = '/home/idies/workspace/Temporary/raddick/census_scratch/acs5/'
    yeardir = basedir + '{0:.0f}/'.format(thisyear)
    rawdatadir = yeardir + 'rawdata/'

    if (thisyear >= 2018):
        metadir = yeardir + 'metadata/'
    elif (thisyear == 2017):
        metadir = yeardir + 'metadata/xls_temp/'
    elif (thisyear == 2016):
        metadir = yeardir + 'metadata/templates/'
    elif (thisyear == 2015):
        metadir = yeardir + 'metadata/'
    elif (thisyear == 2014):
        metadir = yeardir + 'metadata/seq/'
    elif (thisyear == 2013):
        metadir = yeardir + 'metadata/'
    elif (thisyear == 2012):
        metadir = yeardir + 'metadata/'
    elif (thisyear == 2011):
        metadir = yeardir + 'metadata/'
    elif (thisyear == 2010):
        metadir = yeardir + 'metadata/'

    vardir = yeardir + 'variables/'

    for thisdir in [vardir]:#[datadir, errordir, vardir, geodir]:
        #print(thisdir)
        if not(os.path.exists(thisdir)):
            os.makedirs(thisdir)

    sequences = []

    if (thisyear >= 2017):
        sequences = ['seq'+str(y)+'.xlsx' for y in sorted([int(x[3:x.find('.')]) for x in os.listdir(metadir) if 'seq' in x])]
    else:
        sequences = ['Seq'+str(y)+'.xls' for y in sorted([int(x[3:x.find('.')]) for x in os.listdir(metadir) if 'Seq' in x])]

    metadata_df = pandas.DataFrame()

#    for i in range(1,5):
    for i in range(1,len(sequences)+1):
        if (debug >=1):
            if ((np.mod(i, 10) == 0) | (i == len(sequences))):
                print('\treading metadata for sequence {0:.0f}...'.format(i))
        if (thisyear >= 2017):
            this_seq_metadata_filename = metadir + 'seq{0:.0f}.xlsx'.format(i)
        elif (thisyear == 2011):
            this_seq_metadata_filename = metadir + 'Seq{0:04d}.xls'.format(i)
        else:
            this_seq_metadata_filename = metadir + 'Seq{0:.0f}.xls'.format(i)

        this_seq_metadata_in_cols_df = pandas.read_excel(this_seq_metadata_filename, header=None)#, encoding='utf-8')    
        this_seq_metadata_in_cols_df.columns = this_seq_metadata_in_cols_df.loc[0]
        
        this_seq_metadata_df = this_seq_metadata_in_cols_df.T
        this_seq_metadata_df.columns = ['variable', 'description']    
        this_seq_metadata_df = this_seq_metadata_df.assign(sequence_number = i)
        if (i > 1):
            this_seq_metadata_df = this_seq_metadata_df.drop(['FILEID','FILETYPE','STUSAB','CHARITER','SEQUENCE','LOGRECNO'], axis=0)
        this_seq_metadata_df['sequence_number'] == i
        if (debug >= 2):
            print('\t\tMerging datasets...')
        metadata_df = pandas.concat((metadata_df, this_seq_metadata_df), axis=0)

    metadata_df = metadata_df.reset_index(drop=True)
    metadata_df.index.name = 'rownumber'

    print('   writing outfile...')
    metadata_df.to_csv(vardir+'variables_acs_5yr_all.csv', encoding='utf-8')
e = time.time()
print('Done in {0:,.0f} seconds!'.format(e-s))



# Which variables do we need?

# basedir = '/home/idies/workspace/Temporary/raddick/census_scratch/acs5/'
# yeardir = basedir + '{0:.0f}/'.format(2018)
# vardir = yeardir + 'variables/'

# z = pandas.read_csv(vardir+'variables_acs_5yr_all.csv', encoding='utf-8', index_col='rownumber')#.loc[0:80]['description'].tolist()
# # z[
# #     z['variable'].apply(lambda x: 'B01001' in x)
# #     #&  (z['description'].apply(lambda x: 'education' in str(x).lower()))
# #     #& (z['description'].apply(lambda x: '25' in str(x).lower()))
# #     #& (z['sequence_number'] 
# # ]#[46:]
# z

# s = time.time()

# debug = 1

# basedir = '/home/idies/workspace/Temporary/raddick/census_scratch/acs5/'
# yeardir = basedir + '{0:.0f}/'.format(2018)
# rawdatadir = yeardir + 'rawdata/'


# metadata_df[
#     (metadata_df['description'].apply(lambda x: 'education' in str(x).lower()))
#     & (metadata_df['sequence_number'] < 43)
# ]
# #[0:20]
# #metadata_df[metadata_df['variable'] == 'B01001_001']



# # # SEQUENCE 4: RACE
# # #['B02001_001', 'B02001_002', 'B02001_003', 'B02009_001'] 
# # # SEQUENCE 23: TRANSPORTATION
# # # ['B08013_001']
# # # SEQUENCE 36: HOUSEHOLDS
# # #['B11001_001']
# # # SEQUENCE 43: EDUCATION
# # #['B15002_002' to 'B15002_035'; 'C15002A_002' to 'C15002A_011'; 'C15002B_002' to 'C15002B_011']
# # # SEQUENCE 48: POVERTY
# # # ['B17001_001', 'B17001_002', 'B17001A_001', 'B17001A_002', 'B17001B_001', 'B17001B_002']
# # # SEQUENCE 59: INCOME
# # #['B19013_001', 'B19013A_001', 'B19013B_001']
# # # SEQUENCE 63: INCOME
# # #['B19062_001']
# # # SEQUENCE 112: HOUSING
# # #['B25002_001', 'B25002_002', 'B25002_003', 'B25003_001', 'B25003_002', 'B25003_003', 'B25003A_001', 'B25003A_002', 'B25003A_003', 'B25003B_001', 'B25003B_002', 'B25003B_003']
# # # SEQUENCE 115: HOUSING
# # #['B25077_001']

# # sequence_list = [4, 23, 43, 48, 59, 63, 112, 115]
# # for_cra_analysis_mac_varlist = ['B02001_001', 'B02001_002', 'B02001_003', 'B02009_001']
# # for_cra_analysis_mac_varlist += ['B08013_001']
# # for i in range(2, 36):
# #     for_cra_analysis_mac_varlist.append('B15002_{:03d}'.format(i))
# # for i in range(2,12):
# #     for_cra_analysis_mac_varlist.append('C15002A_{:03d}'.format(i))
# # for i in range(2,12):
# #     for_cra_analysis_mac_varlist.append('C15002B_{:03d}'.format(i))
# # for_cra_analysis_mac_varlist += ['B17001_001', 'B17001_002', 'B17001A_001', 'B17001A_002', 'B17001B_001', 'B17001B_002']
# # for_cra_analysis_mac_varlist += ['B19013_001', 'B19013A_001', 'B19013B_001']
# # for_cra_analysis_mac_varlist += ['B19062_001']
# # for_cra_analysis_mac_varlist += ['B25002_001', 'B25002_002', 'B25002_003', 'B25003_001', 'B25003_002', 'B25003_003', 'B25003A_001', 'B25003A_002', 'B25003A_003', 'B25003B_001', 'B25003B_002', 'B25003B_003']
# # for_cra_analysis_mac_varlist += ['B25077_001']
# # # metadata_df[
# # #     (metadata_df['sequence_number'] == 115) & 
# # #     (metadata_df['description'].apply(lambda x: 'value' in x.lower())) & 
# # #     (metadata_df['description'].apply(lambda x: 'median' in x.lower()))
# # # ]#[0:20]
# # #for_cra_analysis_mac_varlist

# def find_variables4(mdf, tofindlist):
#     # NEED TO REWRITE TO USE REGEX INSTEAD, because now only "Male" returns men; "male" returns women b/c feMALE
# #    varlist = []
#     if (len(tofindlist) == 0):
#         print('CAUTION: No search phrases specified, returning every variable!')
#     if (len(tofindlist) > 4):
#         print('ERROR: list contains more than four search phrases, searching for only the first three')
#     try:
#         string1 = tofindlist[0]
#     except IndexError:
#         string1 = ''
#     try:
#         string2 = tofindlist[1]
#     except IndexError:
#         string2 = ''
#     try:
#         string3 = tofindlist[2]
#     except IndexError:
#         string3 = ''
#     try:
#         string4 = tofindlist[3]
#     except IndexError:
#         string4 = ''

# #    for x in tofindlist:
# #        print(x)
        
#     print('')
    
#     varlist = mdf['variable'][mdf['description'].apply(lambda x: (string1 in x) & (string2 in x) & (string3 in x) & (string4 in x))].values.tolist()
    
#     return varlist
# print('Functions defined')

# phrases = []
# #phrases.append('Median household income')
# phrases.append('overty')
# #phrases.append('chool')
# #phrases.append('egree')

# metadata_df['variable'] = metadata_df['variable'].astype('str')
# metadata_df['description'] = metadata_df['description'].astype('str')

# thelist = find_variables4(metadata_df, phrases)
# print('Variables found: ',thelist)

# #metadata_df.dtypes
# metadata_df.loc[thelist]

# #Total population: sequence 2, variable B01001_001
# #Aggregate travel time to work: sequence 23, variable B08013_001
# #White alone: sequence 4, variable B02001_003
# #Black alone: sequence 4, variable B02001_004
# #Two or more races: sequence 4, variable B02001_008
# #Hispanic or Latino origin: sequence 5, variable B03001_003
# #Sex by age by educational attainment: sequence 43, variable B15001_001
# #Poverty status in the past 12 months by sex by age: sequence 48, variable B17001_001
# #Median household income: sequence 59, variable B19013_001
# #Aggregate wage or salary income in past 12 months: sequence 63, variable B19062_001

# #SO, the sequences we need are: 2, 4, 5, 23, 43, 48, 59, 63
# #The variables we need are: B01001_001, B08013_001, B02001_003, B02001_004, B02001_008, B03001_003, B15001_001, B17001_001, B19013_001, B19062_001

#metadata_df[metadata_df['sequence_number'] == 48]
