==> 2012to2017/4. Get ACS5 data for tracts.ipynb <==
#!pip install --upgrade pip
#!pip install numpy
#!pip install pandas
#!pip install xlrd
debug = 1
import numpy as np
import pandas
import time
import geopandas
import os
from pprint import pprint
pandas.set_option('display.max_colwidth', None)
g = 0
print('ok')

# Process data for one year

## Read variable names and descriptions

s = time.time()
thisyear = 2010
# 2019: Wrote out 73,057 tracts and 217,740 block groups in 30 minutes 30 seconds!
# 2018: Wrote out 73,056 tracts and 217,739 block groups in 38 minutes 24 seconds!
# 2017: Wrote out 73,056 tracts and 217,739 block groups in 36 minutes 38 seconds!
# 2016: Wrote out 73,056 tracts and 217,739 block groups in 38 minutes 14 seconds!
# 2015: Wrote out 73,056 tracts and 217,739 block groups in 36 minutes 11 seconds!
# 2014: Wrote out 73,056 tracts and 217,739 block groups in 35 minutes 2 seconds!
# 2013: Wrote out 73,056 tracts and 217,739 block groups in 36 minutes 48 seconds!
# 2012: Wrote out 73,056 tracts and 217,739 block groups in 34 minutes 32 seconds!
# 2011: Wrote out 73,056 tracts and 217,739 block groups in 29 minutes 57 seconds!
# 2010: Wrote out 73,057 tracts and 217,740 block groups in 36 minutes 33 seconds!

category = 'for_cra_analysis_mac'

basedir = '/home/idies/workspace/Temporary/raddick/census_scratch/acs5/'
yeardir = basedir + str(thisyear) + '/'

if (thisyear == 2016):
    rawdatadir = yeardir + 'rawdata/data/tab4/sumfile/prod/2012thru2016/group2/'
elif (thisyear in [2015, 2013, 2011]):
    rawdatadir = yeardir + 'rawdata/group2/'
elif (thisyear == 2014):
    rawdatadir = yeardir + 'rawdata/tab4/sumfile/prod/2010thru2014/group2/'
elif (thisyear == 2012):
    rawdatadir = yeardir + 'rawdata/tab4/sumfile/prod/2008thru2012/group2/'
else:
    rawdatadir = yeardir + 'rawdata/'
metadir = yeardir + 'metadata/'
vardir = yeardir + 'variables/'
geodir = yeardir + 'geography/'
estimates_dir = yeardir + 'estimates/'
margins_of_error_dir = yeardir + 'margins_of_error/'

for x in [estimates_dir, margins_of_error_dir]:
    if not(os.path.exists(x)):
        os.makedirs(x)
    if not(os.path.exists(x+'bg/')):
        os.makedirs(x+'bg/')

this_dir = '/home/idies/workspace/Storage/raddick/Baltimore/community_reinvestment_act/'
extras_dir = '/home/idies/workspace/Storage/raddick/census/extras/'

all_variables_df = pandas.read_csv(vardir+'variables_acs_5yr_all.csv', low_memory=False, index_col='rownumber')

# pre-2017, description is blank for FILEID, FILETYPE, STUSAB, CHARITER, SEQUENCE, LOGRECNO - so fill in with blank
all_variables_df = all_variables_df.fillna('')

print('Done')

#all_variables_df

## Figure out which variables and sequences we need for our work

See the code at the bottom of this file for how I got all the variables we are getting.

# all_variables_df[all_variables_df['description'].apply(lambda x: 'vacant' in x.lower())][0:10]

# # for i in range(11,50):
# #     print('\'B01001_{0:03d}\','.format(i))

# all_variables_df[all_variables_df['variable'].isin(['B25002_001','B25002_002','B25002_003'])]

# Which variables are we getting?
for_cra_analysis_mac_varlist = ['FILEID',
 'FILETYPE',
 'STUSAB',
 'CHARITER',
 'SEQUENCE',
 'LOGRECNO',
# 'B00001_001',
# 'B00002_001',
 'B01001_001',
 'B02001_001',
 'B02001_002',
 'B02001_003',
 'B08013_001',
 'B11001_001',
 'B11001_002',
 'B11001_006',
 'B11001_007',
 'B11001A_001', 
 'B11001B_001',
'B01001_011',
'B01001_012',
'B01001_013',
'B01001_014',
'B01001_015',
'B01001_016',
'B01001_017',
'B01001_018',
'B01001_019',
'B01001_020',
'B01001_021',
'B01001_022',
'B01001_023',
'B01001_024',
'B01001_025',
'B01001_026',
'B01001_027',
'B01001_028',
'B01001_029',
'B01001_030',
'B01001_031',
'B01001_032',
'B01001_033',
'B01001_034',
'B01001_035',
'B01001_036',
'B01001_037',
'B01001_038',
'B01001_039',
'B01001_040',
'B01001_041',
'B01001_042',
'B01001_043',
'B01001_044',
'B01001_045',
'B01001_046',
'B01001_047',
'B01001_048',
'B01001_049',                                                                
 'B15003_001',
 'B15003_002',
 'B15003_003',
 'B15003_004',
 'B15003_005',
 'B15003_006',
 'B15003_007',
 'B15003_008',
 'B15003_009',
 'B15003_010',
 'B15003_011',
 'B15003_012',
 'B15003_013',
 'B15003_014',
 'B15003_015',
 'B15003_016',
 'B15003_017',
 'B15003_018',
 'B15003_019',
 'B15003_020',
 'B15003_021',
 'B15003_022',
 'B15003_023',
 'B15003_024',
 'B15003_025',
 'B17001_001',
 'B17001_002',
 'B17001A_001',
 'B17001A_002',
 'B17001B_001',
 'B17001B_002',
 'B19013_001',
 'B19013A_001',
 'B19013B_001',
 'B19113_001',
 'B23025_002',
 'B23025_005',
 'B25002_001',
 'B25002_002',
 'B25002_003',
 'B25003_001',
 'B25003_002',
 'B25003_003',
 'B25003A_001',
 'B25003A_002',
 'B25003A_003',
 'B25003B_001',
 'B25003B_002',
 'B25003B_003',
 'B25077_001',
 'B25034_001',
 'B25035_001']

# all_variables_df[all_variables_df['variable'].isin(for_cra_analysis_mac_varlist)]['description'][5:50]

# Which sequences will we need?
sequence_list = all_variables_df[all_variables_df['variable'].isin(for_cra_analysis_mac_varlist)]['sequence_number'].drop_duplicates().tolist()
e = time.time()
g = g + (e-s)

sequence_list


#for_cra_analysis_mac_varlist



s = time.time()

state_codes_df = pandas.read_csv(extras_dir+'statecodes.csv')
state_codes_df = state_codes_df.set_index('STUSAB')

estimates_df = pandas.DataFrame()
margins_of_error_df = pandas.DataFrame()
states = state_codes_df.index.values.tolist()
states = [x.lower() for x in states if x not in ('AS', 'GU', 'MP', 'PR', 'UM', 'VI')]

for i in sequence_list:
    if (debug >= 1):
        print('Importing estimates and margins of error for sequence {0:,.0f}...'.format(i))
    #print(all_variables_df[all_variables_df['sequence_number'] == i]['variable'].tolist()[0:10])
    
    this_seq_estimates_df = pandas.DataFrame()
    this_seq_margins_of_error_df = pandas.DataFrame()
    for onestate in states:
#        if (debug >= 2):
#            print('\tImporting estimates and margins of error for {0:}...'.format(onestate))
        state_estimates_filename = rawdatadir + 'e{0:.0f}5{1:}{2:04d}000.txt'.format(thisyear, onestate, i)
        state_margins_of_error_filename = rawdatadir + 'm{0:.0f}5{1:}{2:04d}000.txt'.format(thisyear, onestate, i)
#         if ('e{0:.0f}5{1:}{2:04d}000.txt'.format(thisyear, onestate, i) not in os.listdir(rawdatadir)):
#             print('{0:} not found!'.format(state_estimates_filename))

        onestate_estiamtes_df = pandas.read_csv(state_estimates_filename, header=None, sep=',', encoding='utf-8', low_memory=False)
        this_seq_estimates_df = this_seq_estimates_df.append(onestate_estiamtes_df)
        onestate_margins_of_error_df = pandas.read_csv(state_margins_of_error_filename, header=None, sep=',', encoding='utf-8', low_memory=False)
        this_seq_margins_of_error_df = this_seq_margins_of_error_df.append(onestate_margins_of_error_df)
    if (i >= 2):
        #print(this_seq_estimates_df.columns.tolist())
        this_seq_estimates_df = this_seq_estimates_df.drop([0,1,2,3,4,5], axis=1)  # ['FILEID','FILETYPE','STUSAB', 'SEQUENCE', 'CHARITER','LOGRECNO']
        this_seq_margins_of_error_df = this_seq_margins_of_error_df.drop([0,1,2,3,4,5], axis=1)  # ['FILEID','FILETYPE','STUSAB', 'SEQUENCE', 'CHARITER','LOGRECNO']

    this_seq_estimates_df.columns = all_variables_df[all_variables_df['sequence_number'] == i]['variable'].tolist()
    this_seq_margins_of_error_df.columns = all_variables_df[all_variables_df['sequence_number'] == i]['variable'].tolist()
    
    estimates_df = pandas.concat((estimates_df,this_seq_estimates_df), axis=1, sort=False)
    margins_of_error_df = pandas.concat((margins_of_error_df,this_seq_margins_of_error_df), axis=1, sort=False)
    
print('checking whether the dataset has the variables we need...')

#requested_variables = all_variables_df[all_variables_df['sequence_number'] == 1]['variable'].tolist()
requested_variables = for_cra_analysis_mac_varlist

variables_to_get = []
variables_not_found = []
for x in requested_variables:
    if x in estimates_df.columns:
        variables_to_get.append(x)
    else:
        variables_not_found.append(x)

        
# if (len(variables_not_found) > 0):
#     print('Variables not found:')
#     for y in variables_not_found:
#         print('{0:}: {1:}\n'.format(y, all_variables_df[all_variables_df['variable'] == y]['description'].values[0]))

print('keeping only the variables we need...')
estimates_df = estimates_df[variables_to_get]
margins_of_error_df = margins_of_error_df[variables_to_get]

#all_variables_df[all_variables_df['variable'].isin(variables_not_found)]['description']

print('backing up...')
estimates_df_bk = estimates_df
margins_of_error_df_bk = margins_of_error_df

e = time.time()
g = g + (e-s)

print('\nFound estimates for {0:,.0f} tracts and margins of error for {1:,.0f} tracts in {2:,.0f} minutes {3:,.0f} seconds!'.format(len(estimates_df), len(margins_of_error_df), np.floor((e-s)/60), np.floor((e-s)%60)))


#estimates_df['STUSAB']

## Read geographies

s = time.time()
debug = 1

geo_df = pandas.DataFrame()

states = state_codes_df.index.values.tolist()
states = [x.lower() for x in states if x not in ('AS', 'GU', 'MP', 'PR', 'UM', 'VI')]

if (debug >= 1):
    print('Reading geography files...')
for onestate in states:
    if (thisyear >= 2016):
        filename = geodir+'{0:}.xlsx'.format(onestate)
    else:
        filename = geodir+'{0:}.xls'.format(onestate)
    if (debug == 2):
        print('Reading geography for {0:}...'.format(onestate))
    this_geo_df = pandas.read_excel(filename)
    geo_df = pandas.concat((geo_df, this_geo_df), sort=False)
    if (thisyear in [2017, 2018]):
        geo_df = geo_df.assign(STUSAB = geo_df['State'].apply(lambda x: x.lower()))
    else:
        geo_df = geo_df.assign(STUSAB = geo_df['STATE'].apply(lambda x: x.lower()))

geo_df = geo_df.rename(columns={'Geography ID': 'GEOID'})

print('Retaining only low-level geographies...')
geo_tracts_df = geo_df[(geo_df['GEOID'].apply(lambda x: x[0:3] == '140'))]
if (debug >= 1):
    print('Retained {0:,.0f} tract level geographies...'.format(len(geo_tracts_df)))

geo_bg_df = geo_df[(geo_df['GEOID'].apply(lambda x: x[0:3] == '150'))]
if (debug >= 1):
    print('Retained {0:,.0f} tract level geographies...'.format(len(geo_bg_df)))

print('backing up...')
geo_tracts_df_bk = geo_tracts_df
geo_bg_df_bk = geo_bg_df

##geo_df = geo_df.set_index('GEOID')  #We'll set GEOID as index colum AFTER the merge

e = time.time()
g = g + (e-s)
print('Done in {0:.0f} minutes {1:,.0f} seconds!'.format( np.floor((e-s)/60), (e-s) % 60))

#geo_df.sample(1)
#print('skipping...')
geo_tracts_df.head(1).T

## Join geography to estimates and margins of error

s = time.time()
print('getting from backup...')
estimates_df = estimates_df_bk
margins_of_error_df = margins_of_error_df_bk
geo_tracts_df = geo_tracts_df_bk
geo_bg_df = geo_bg_df_bk

#.sort_values('Logical Record Number')
#estimates_df[['STUSAB','LOGRECNO']].dtypes #object, int64
#geo_df[['STUSAB', 'Logical Record Number']].dtypes

if (thisyear in [2017, 2018]):
    estimates_tracts_df = estimates_df.merge(geo_tracts_df, left_on=['STUSAB', 'LOGRECNO'], right_on=['STUSAB', 'Logical Record Number'])
    margins_of_error_tracts_df = margins_of_error_df.merge(geo_tracts_df, left_on=['STUSAB', 'LOGRECNO'], right_on=['STUSAB', 'Logical Record Number'])

    estimates_bg_df = estimates_df.merge(geo_bg_df, left_on=['STUSAB', 'LOGRECNO'], right_on=['STUSAB', 'Logical Record Number'])
    margins_of_error_bg_df = margins_of_error_df.merge(geo_bg_df, left_on=['STUSAB', 'LOGRECNO'], right_on=['STUSAB', 'Logical Record Number'])
    
# elif (thisyear == 2016):
#     estimates_df = estimates_df.merge(geo_df, left_on=['STUSAB', 'LOGRECNO'], right_on=['STUSAB', 'Logical Record Number'])
#     margins_of_error_df = margins_of_error_df.merge(geo_df, left_on=['STUSAB', 'LOGRECNO'], right_on=['STUSAB', 'LOGRECNO'])
else:
    estimates_tracts_df = estimates_df.merge(geo_tracts_df, left_on=['STUSAB', 'LOGRECNO'], right_on=['STUSAB', 'LOGRECNO'])
    margins_of_error_tracts_df = margins_of_error_df.merge(geo_tracts_df, left_on=['STUSAB', 'LOGRECNO'], right_on=['STUSAB', 'LOGRECNO'])

    estimates_bg_df = estimates_df.merge(geo_bg_df, left_on=['STUSAB', 'LOGRECNO'], right_on=['STUSAB', 'LOGRECNO'])
    margins_of_error_bg_df = margins_of_error_df.merge(geo_bg_df, left_on=['STUSAB', 'LOGRECNO'], right_on=['STUSAB', 'LOGRECNO'])

estimates_tracts_df = estimates_tracts_df.set_index('GEOID')  # set the GEOID after we add the shapefiles
margins_of_error_tracts_df = margins_of_error_tracts_df.set_index('GEOID')  # set the GEOID after we add the shapefiles
estimates_bg_df = estimates_bg_df.set_index('GEOID')
margins_of_error_bg_df = margins_of_error_bg_df.set_index('GEOID')

print('Total estimate by tract of B01001_001: {0:,.0f}'.format(estimates_tracts_df['B01001_001'].sum()))
print('Margin of error by tract in B01001_001: {0:,.0f}'.format(margins_of_error_tracts_df['B01001_001'].sum()))

print('Total estimate by block group of B01001_001: {0:,.0f}'.format(estimates_bg_df['B01001_001'].sum()))
print('Margin of error by tract in B01001_001: {0:,.0f}'.format(margins_of_error_bg_df['B01001_001'].sum()))

print('\n')

print('Writing data...')

estimates_tracts_df.to_csv(estimates_dir + 'estimates_acs{0:.0f}_tract_{1:}.csv'.format(thisyear, category), encoding='utf-8')
print('Estimates by tract written to estimates_acs{0:.0f}_tract_{1:}.csv '.format(thisyear, category))
      
margins_of_error_tracts_df.to_csv(margins_of_error_dir + 'margins_of_error_acs{0:.0f}_tract_{1:}.csv'.format(thisyear, category), encoding='utf-8')
print('Margins of error by tract written to margin_of_error_acs{0:.0f}_tract_{1:}.csv'.format(thisyear, category))

estimates_bg_df.to_csv(estimates_dir + 'bg/estimates_acs{0:.0f}_bg_{1:}.csv'.format(thisyear, category), encoding='utf-8')
print('Estimates by block group written to bg/estimates_acs{0:.0f}_bg_{1:}.csv '.format(thisyear, category))
      
margins_of_error_bg_df.to_csv(margins_of_error_dir + 'bg/margins_of_error_acs{0:.0f}_bg_{1:}.csv'.format(thisyear, category), encoding='utf-8')
print('Margins of error by block group written to bg/margin_of_error_acs{0:.0f}_bg_{1:}.csv'.format(thisyear, category))

e = time.time()
g = g + (e-s)
print('\nWrote out {0:,.0f} tracts and {1:,.0f} block groups in {2:,.0f} minutes {3:,.0f} seconds!'.format(len(estimates_tracts_df), len(estimates_bg_df), np.floor((g)/60), np.floor((g)%60)))

#margins_of_error_df

z = pandas.read_csv(estimates_dir+'estimates_acs{0:}_tract_for_cra_analysis_mac.csv'.format(thisyear), encoding='utf-8', low_memory=False, index_col='GEOID')
z

# # thevars = pandas.
# z.columns.tolist()
# vars_df = pandas.read_csv(vardir+'variables_acs_5yr_all.csv', index_col='rownumber')

# vars_df[
#     vars_df['variable'].isin(z.columns.tolist())][['variable', 'description']
#                                                  ].to_csv(this_dir+'cra_study_variables.csv')

# print('Done!')
